# Optimize block map construction and multiplication