--- a/src/blockmap.jl
+++ b/src/blockmap.jl
@@ -11,7 +11,7 @@ end
 
 BlockMap{T}(maps::As, rows::S) where {T,As<:Tuple{Vararg{LinearMap}},S} = BlockMap{T,As,S}(maps, rows)
 
-firstindices(maps::Tuple{Vararg{LinearMap}}, dim) = cumsum([1, map(m -> size(m, dim), maps)...,])
+# firstindices(maps::Tuple{Vararg{LinearMap}}, dim) = cumsum([1, map(m -> size(m, dim), maps)...,])
 
 function check_dims(maps::Tuple{Vararg{LinearMap}}, k)
     n = size(maps[1], k)
@@ -20,6 +20,22 @@ function check_dims(maps::Tuple{Vararg{LinearMap}}, k)
     end
     return nothing
 end
+
+sumshift(a::Tuple{Vararg{Int}}) = sumshift((1,), a)
+sumshift(a::Tuple{Vararg{Int}}, as::Tuple{Vararg{Int}}) = sumshift(tuple(a..., last(a) + first(as)), Base.tail(as))
+sumshift(a::Tuple{Vararg{Int}}, b::Tuple{Int}) = tuple(a..., last(a) + first(b))
+
+function firstrowindices(A::BlockMap)
+    as, rows = A.maps, A.rows
+
+	firstrows = sumshift(A.rows)
+	Afirstcol = ntuple(length(firstrows)-1) do i
+		A.maps[firstrows[i]]
+	end
+	return sumshift(map(a -> size(a, 1), Afirstcol))
+end
+
+@inline firstcolindices(As::Tuple{Vararg{LinearMap}}) = sumshift(map(a -> size(a, 2), As))
 
 function Base.size(A::BlockMap)
     as, rows = A.maps, A.rows
@@ -205,9 +221,9 @@ LinearAlgebra.adjoint(A::BlockMap)  = AdjointMap(A)
 function A_mul_B!(y::AbstractVector, A::BlockMap, x::AbstractVector)
     maps, rows = A.maps, A.rows
     mapind = 0
-    yinds = firstindices(maps[cumsum([1, rows...])[1:end-1]], 1)
+    yinds = firstrowindices(A)
     @views for rowind in 1:length(rows)
-        xinds = firstindices(maps[mapind+1:mapind+rows[rowind]], 2)
+        xinds = firstcolindices(maps[mapind+1:mapind+rows[rowind]])
         yrow = @views y[yinds[rowind]:(yinds[rowind+1]-1)]
         mapind += 1
         A_mul_B!(yrow, maps[mapind], x[xinds[1]:xinds[2]-1])
@@ -223,9 +239,9 @@ function At_mul_B!(y::AbstractVector, A::BlockMap, x::AbstractVector)
     maps, rows = A.maps, A.rows
     fill!(y, 0)
     mapind = 0
-    xinds = firstindices(maps[cumsum([1, rows...])[1:end-1]], 1)
+    xinds = firstrowindices(A)
     # first block row (rowind = 1), fill all of y
-    yinds = firstindices(maps[mapind+1:mapind+rows[1]], 2)
+    yinds = firstcolindices(maps[mapind+1:mapind+rows[1]])
     xcol = @views x[xinds[1]:(xinds[2]-1)]
     @views for colind in 1:rows[1]
         mapind +=1
@@ -233,7 +249,7 @@ function At_mul_B!(y::AbstractVector, A::BlockMap, x::AbstractVector)
     end
     # subsequent block rows, add results to corresponding parts of y
     @views for rowind in 2:length(rows)
-        yinds = firstindices(maps[mapind+1:mapind+rows[rowind]], 2)
+        yinds = firstcolindices(maps[mapind+1:mapind+rows[rowind]])
         xcol = @views x[xinds[rowind]:(xinds[rowind+1]-1)]
         for colind in 1:rows[rowind]
             mapind +=1
@@ -247,9 +263,9 @@ function Ac_mul_B!(y::AbstractVector, A::BlockMap, x::AbstractVector)
     maps, rows = A.maps, A.rows
     fill!(y, 0)
     mapind = 0
-    xinds = firstindices(maps[cumsum([1, rows...])[1:end-1]], 1)
+    xinds = firstrowindices(A)
     # first block row (rowind = 1), fill all of y
-    yinds = firstindices(maps[mapind+1:mapind+rows[1]], 2)
+    yinds = firstcolindices(maps[mapind+1:mapind+rows[1]])
     xcol = @views x[xinds[1]:(xinds[2]-1)]
     @views for colind in 1:rows[1]
         mapind +=1
@@ -257,7 +273,7 @@ function Ac_mul_B!(y::AbstractVector, A::BlockMap, x::AbstractVector)
     end
     # subsequent block rows, add results to corresponding parts of y
     @views for rowind in 2:length(rows)
-        yinds = firstindices(maps[mapind+1:mapind+rows[rowind]], 2)
+        yinds = firstcolindices(maps[mapind+1:mapind+rows[rowind]])
         xcol = @views x[xinds[rowind]:(xinds[rowind+1]-1)]
         for colind in 1:rows[rowind]
             mapind +=1