I couldn't stop myself from further optimizing things, partly for my own education (@JeffFessler 😉 ). Let me start with a simple "benchmark problem".
```julia
using LinearMaps, LinearAlgebra, BenchmarkTools
A = LinearMap(rand(10, 10))
L = hvcat((2, 2), A, A, A, I)
x = rand(size(L, 2))
y = rand(size(L, 1))
@btime mul!($y, $L, $x)
M = Matrix(L)
@btime mul!($y, $M, $x)
```
On current master, we get:
```julia
 2.682 μs (33 allocations: 2.23 KiB)
  97.825 ns (0 allocations: 0 bytes)
```
On this branch, we get:
```julia
 1.365 μs (17 allocations: 896 bytes)
  96.926 ns (0 allocations: 0 bytes)
```
So, a factor of 2 on this small problem. What has happened? I changed from the `Vector`-based `cumsum` formulation of obtaining the starting indices corresponding to blocks, and changed it to a tuple-based, recursive, inferable formulation. Obviously, the values obtained are the exact same in all cases. This is responsible for the major improvement above. This is compiler magic that I haven't expected myself.

What else? We are allocating memory because we generally can't compute the target vectors component-wise. So when we add the result of some block multiplication to a previous result, we need to first allocate the intermediate result before we can add it. This is, in general, unavoidable. BUT, there are two exceptions: uniform scalings (for which I have an allocation-free (this is tested!), 5-arg `mul!`), and wrapped maps with BLAS-matrices. Unfortunately, we don't have that 5-arg mul! for (dense) matrices, yet. But the minute that is coming, our code will perform better immediately. The remaining allocations are likely to generate views. For instance
```julia
L = [A I I]
x = rand(size(L, 2))
y = rand(size(L, 1))
@btime mul!($y, $L, $x)
```
should not be allocating at all, because the identities are handled by the allocation-free 5-arg `mul!`, but we still get
```julia
 687.351 ns (14 allocations: 624 bytes)
```
in comparison to
```julia
M = Matrix(L)
@btime mul!($y, $M, $x)
133.692 ns (0 allocations: 0 bytes)
```
In good cases, I think the overhead is pretty much constant, and is likely to be irrelevant when costly matrix multiplication is involved.

Other minor improvements include that size checks are now done at linear map promotion.