@dkarrasch 2019-07-27
I wasn't aware of `Dims`, that looks indeed as what we need. Thanks!

The `cumsum` thing is actually nothing but recursion, no iterator. I call it `sumshift` to indicate that in each recursion step, you shift the first element from the second tuple over to the first, while adding it the last entry of the first tuple. Since at compile time, you know the length of the tuples and their element types, the type of the result is perfectly type inferable, and the program will know what to expect, except for the precise values.

I guess (but as I said, I was surprised myself), that this allows for subsequent optimizations by the compiler, which reduces much of the overhead. Perhaps, in Matlab you wouldn't gain anything over plain cumsum, so that's why nobody would care to write a recursive function in that place.