@dkarrasch 2019-07-27
Yes, `Dims` works, but I wonder if it creates a bit of confusion. It is meant to "represent the dimensions of an `AbstractArray`.", as the docs say. So, in a `Matrix`/`LinearMap` context, you would expect to see two-tuples, and not higher-order tuples, which represent the first index of blocks...