@dkarrasch 2019-07-28
Hey, look at this!
```julia
using LinearMaps, LinearAlgebra, BenchmarkTools
A = LinearMap(rand(100, 100))
L = hvcat((3, 3), A, I, I, A, I, I)
x = rand(size(L, 2))
y = rand(size(L, 1))
@btime mul!($y, $L, $x)
M = Matrix(L)
@btime mul!($y, $M, $x)
```
gives
```julia
6.269 μs (32 allocations: 1.34 KiB)
7.166 μs (0 allocations: 0 bytes)
```
The larger the blocks, the more it pays off to have the optimized uniform scaling multiplication and to really use the block structure. In that sense, this the clearly the way to go, and not to put everything in one big matrix via `BlockArrays.jl`.

Generally, there is probably still much room for improvement. In the multiplication, there are the two calls to `firstrowindices` and to `firstcolindices`. Both are inferable, but allocate. I'm having fun trying to optimize this, but couldn't find a better solution yet. Attempts like
```julia
function firstrowindices(A::LinearMaps.BlockMap)
    as, rows = A.maps, A.rows

    index = 1
    cumsize = 1
    return tuple(1, ntuple(length(rows)-1) do i
        cumsize += size(as[index], 1)
        index += rows[i]
        return cumsize
    end...)
end
```
and
```julia
sumshift(a::Tuple{Vararg{Int}}, b::Int) = tuple(a..., last(a) + b)
colindices(As::Tuple{Vararg{LinearMap}}) = mapreduce(a -> size(a, 2), sumshift, As; init=(1,))
```
don't seem to improve anything. Anyway, this is in pretty good shape already, and I'm excited if we manage to further improve it.