@dkarrasch 2019-07-28
Haha, I just found @Jutho's TupleTools.jl and the cumsum function there. Maybe we could use some of those functions (I like the `getindex` for instance), but my first attempts again don't improve anything. I guess I'll call it a day on a Sunday, but remain excited if @Jutho can squeeze out some more performance.