@dkarrasch 2019-07-29
Maybe it's as easy as that: compute the block indices at construction and save them in a field of the `BlockMap` struct? There is no reason to repeat this computation at every multiplication, because this is tied to the BlockMap. This should take off lots of the pressure to optimize out these steps.