@dkarrasch 2019-07-29
Alright, in my little 10x10 benchmark from the OP, I am now down on this:
```julia
422.231 ns (8 allocations: 480 bytes)
  97.380 ns (0 allocations: 0 bytes)
```
This is faster by almost factor 8 than current master. I figured that since now the range computations are one-time computations, there is little if any benefit from working with tuples and recursive code. Instead, I use nice concrete `Vector`s of either `Int`s or `UnitRange{Int}`. I have peppered the multiplication with `@inbounds` after checking that the total size of `x` and `y` match the corresponding dimensions of `A`. The internal subindexing is then "known" to be correct because this is how we computed the index ranges at construction of `A`.

I believe the code is also pretty clean by now. What is still missing is an attempt to automatically identify symmetry, and potentially some nice `show` method.