@dkarrasch 2019-07-30
Yes, I'm almost done with the symmetry stuff, then I'll let go.