@dkarrasch 2019-07-30
After tests pass, this is ready for review. To summarize:

The new philosophy is to compute all ranges at construction. This reduces the multiplication overhead as described above. I couldn't get anything better with tuples and recursion here. Early on, I moved the dimension checks into the promotion part, as @Jutho suggested in the first PR. Then, symmetry checks now work, at least for block symmetric/Hermitian operators. So this should probably get a comment like "sufficient condition" or so. Finally, I have added a 5-arg `mul!` for `UniformScalingMap`s, to avoid any allocation there.