@JeffFessler 2019-07-30
I agree that precomputing the ranges at construction with simple code is the way to go.
I should have thought of suggesting this earlier because I do something analogous (but messier) in my matlab toolbox
`https://github.com/JeffFessler/mirt/blob/master/systems/%40fatrix2/private/fatrix2_vertcat.m`