@Jutho 2019-08-08
Looks great; a few suggestions, but otherwise ready to merge I think?