@dkarrasch 2019-08-08
Many thanks to all for reviewing and approving. @Jutho If I don't get to it by tomorrow, please feel free to take over and push it over the finish line. I'll be offline for more than 2 weeks then.