@dkarrasch 2019-08-09
Alright, I think this PR now has everything: updated README, updated version, tests on all Julia versions that are maintained, and all issues/comments resolved. I'll wait a few minutes for all tests to pass (they do pass locally), then merge, then open a registrator issue.