@Jutho 2019-08-09
One question / idea: With the new way of dealing with `hcat` and friends, we could actually also support `AbstractMatrix` types which are not wrapped as `LinearMap`. As soon as there is one `LinearMap` in the list, our methods would take over. But probably this needs a small adaptation to then wrap these arguments as LinearMap, I think only in the `promote_to_lmaps` step.