@Jutho 2019-08-09
But maybe that's an idea for a next PR. Anyway, this is definitely good to merge in my opinion.