@dkarrasch 2019-08-09
Yes, that was my very early dream, about which I have forgotten with all the struggle in between. Wouldn't it simply require
```julia
for k in 1:8 # is 8 sufficient?
      Is = ntuple(n->:($(Symbol(:A,n))::Union{UniformScaling,AbstractMatrix}), Val(k-1))
      L = :($(Symbol(:A,k))::LinearMap)
      args = ntuple(n->Symbol(:A,n), Val(k))

      @eval Base.hcat($(Is...), $L, As::Union{LinearMap,UniformScaling,AbstractMatrix}...) = _hcat($(args...), As...)
      @eval Base.vcat($(Is...), $L, As::Union{LinearMap,UniformScaling,AbstractMatrix}...) = _vcat($(args...), As...)
      @eval Base.hvcat(rows::Tuple{Vararg{Int}}, $(Is...), $L, As::Union{LinearMap,UniformScaling,AbstractMatrix}...) = _hvcat(rows, $(args...), As...)
end
```
? Maybe we should increase the number 8 then, to allow for more than 8 matrices/scalings first...